Rollback Restify 7 Support
In attempting to consume this library in our primary project it was discovered that there was an issue with the restify resolution between the library and the running app.  Since this doesn’t work I’m rolling back support of this feature for now.  This will impact anyone using the asyncConditionalHandler but since it didn’t seem to work I’ll be surprised if anyone is hurt. If you’re using this with Restify 7+ you’re not going to be able to use versioned routes.

This also has some documentation updates and a move to a restify peer dependency since that’s a proper description of the relationship rather than the hard dependency.